+++ src/anteumbra/application/runtime_container.py
@@ -28,6 +28,7 @@ from anteumbra.domain.service_ports import (
 
 if TYPE_CHECKING:
     from anteumbra.application.config_history_service import ConfigHistoryLogger
+    from anteumbra.application.login_rate_service import LoginRateLimiter
     from anteumbra.application.password_service import PasswordService
     from anteumbra.application.quarantine_service import QuarantineService
     from anteumbra.application.scan_state_service import ScanRuntimeState
@@ -48,6 +49,7 @@ class RuntimeContainer:
     passwords: PasswordService
     config_history: ConfigHistoryLogger
     scan_state: ScanRuntimeState
+    login_rate_limiter: LoginRateLimiter
     metrics: RuntimeMetricsPort
     notifier: NotifierPort
     siem_exporter: SIEMExporterPort